Rework storage of ForeignFeeds

+++ app/src/main/java/com/pipe_network/app/domain/entities/Feed.kt
@@ -8,7 +8,7 @@ data class Feed(
     val uuid: String,
     val text: String,
     val friend: Friend,
-    val timestamp: BigInteger,
+    val timestamp: Long,
     val picture: ByteArray? = null,
 ) {
     override fun equals(other: Any?): Boolean {
